Bump versions
The gevent upgrade is not as drastic as it looks. It counts versions
differently now. The bug fixes would be good to get in.

From transifex-client version 0.13.0 they specify a max Python version.
There is an [open PR][1] to support Python 3.10. I hope they fix their
broken build, and get it merged and released soon.

  [1]: https://github.com/transifex/transifex-client/pull/330