Add pickle protocol 5 for older Pythons
The highest protocol on Python 3.6 is 4. The pickle5 library backports protocol 5, which was added in Python 3.8. The tests added here are intended to fail when a new protocol becomes available (for example, when upgrading Python). This is so libraries like django_redis, which uses the highest available protocol by default, are able to read pickles written after upgrading and then reverting due to some unforeseen issue forcing a downgrade.