Refs QA-3629 Refine behavior of matching excel fields to deprecated properties
Do not list deprecated properties in dropdown, but allow them to be typed and auto-suggested
When a deprecated property has been typed/suggested, show a warning that it is a deprecated property
Also show any formats/valid values in this case
Uncheck the "Create new property" in this case as well, since the property exists

In addition, this PR sanitizes allowed values input into the data dictionary before sending to the backend.
(Having script tags (for example) in the backend for suggested valid values causes the importer matching page
to not work as expected -- it seems the js initial data doesn't get deserialied properly -- so stripping in
the same way as is done for display on data dictiony seemed easist to fix this oddity.)

Finally, as requested by Daniel, disable the Next button on the initial upload page when there is no file
chosen.