only add cases to device if they have updates in the form
The previous code assumed that if a case index referenced another
case then that case must be on the device. This is not
necessarily the true (e.g. case search).

This change will only add the case to the device if the case
is also updated / created in the form.