Cache recently used query result namedtuples...
rather than reconstructing for every query. Uses the default lru_cache maxsize of 128, which should be plenty. Currently fetchall_as_namedtuple is used in less than 20 places, and some of those occurrences use the same column names. The cache size can be customized by changing the decorator to `@lru_cache(maxsize=...)` if necessary in the future.

Ref https://rhye.org/post/python-cassandra-namedtuple-performance/