Allow evalute_couch_model_for_sql to be interrupted
Doc types with a very large number of documents can take a long time to process. A random sampling of documents is probably good enough for most cases, and evaluating the first N documents is probably a suitably random sampling since most documents have no meaningful order in the all_docs/by_doc_type view because they  are keyed with random UUIDs.

DocsIterator also improves on the previous solution because it does not load all document ids into memory before any documents can be evaluated. Documents are loaded in chunks of 100 by default, and the chunk size can be customized with a command line parameter.