Add --divided-we-run=XX option to test command
Make it easier to see what options are used by each CI runner. This option is particularly important because it selects a subset of tests, and is often necessary when attempting to track down failures related to bad test cleanup affecting later tests. The option is redundant with the NOSE_DIVIDED_WE_RUN environment variable, which is also set (and must be set for datadog test metrics).