DEFAULT_AUTO_FIELD to avoid deprecation warnings
The default (new in Django 3.2) is BigAutoField, but we have many models that were created with the old default. It is easier to set our global default to `AutoField` since that does not require migrating existing fields. If in the future we decide to change `DEFAULT_AUTO_FIELD` from `AutoField` to `BigAutoField` we have these options for existing models with implicit primary keys:

- Migrate primary key to `BigAutoField`.
- Explicitly add `id = models.AutoField(primary_key=True)` to override the default (also requires a state-only migration).