Simplify Decimal coersion
The old code did a "hack" where it checked for a specific scientific
notation format and rounded to 0 anything with a negative exponent
smaller larger than 3.  It would fail on any other input provided in
scientific notation with a negative exponent - I'm not sure why.
`Decimal()` accepts scientific notation itself, so this change
simplifies that processing considerably, and introduces support for
numbers between 0.001 and -0.001 if provided in scientific notation.