Ignore deprecated "property_match" in tests
"property_match" is a feature of userreports. It is fine that we have tests for a deprecated feature, but those tests should probably be updated to use something like `warnings.catch_warnings()` to ignore the warnings so they do not cause errors or other deprecation output. Updating tests that way is out of scope for the current effort. The whitelist item serves as a future reminder to address this issue.

It also appears to be used in some tests that were not obviously testing that feature alone. Maybe those usages should be changed to use the preferred "boolean_expression" instead?

See corehq/apps/userreports/filters/factory.py