Ignore "TestResult has no addExpectedFailure method"
This warning originates inside of Python's `unittest` framework, and possibly has something to do with `nose`, although I didn't find any evidence of that. On first glance it was not clear what in our test code was triggering it or how to resolve it, but it does appear to cause the intended outcome (test failure is ignored).