Fix BeautifulSoup warning
GuessedAtParserWarning: No parser was explicitly specified, so I'm using the best available HTML parser for this system ("lxml"). This usually isn't a problem, but if you run this code on another system, or in a different virtual environment, it may use a different parser and behave differently.

The code that caused this warning is on line 83 of the file corehq/apps/data_dictionary/tests/test_import_export.py. To get rid of this warning, pass the additional argument 'features="lxml"' to the BeautifulSoup constructor.