Remove "Dispatcher" that isn't
As far as I can tell, this was never used as a dispatcher.  In the
original PR (#1398) it was only used for this purpose.  It is not hooked
up to a URLs file like the other dispatchers.  Rather, it seems to be a
bizarrely indirect way of accessing a constant.