Remove crs custom report
This has been broken for two years.  I commented on that shortly after,
and it looks like we haven't needed to fix it since then.
https://docs.google.com/document/d/18uyJLLFuDZYWDfV6G6MduKF8DspaS0EwH8sK-1gFHsw/edit#heading=h.igds5bmue08
This is the breaking change (`filters.range` doesn't exist - it should
be `filters.date_range`)
https://github.com/dimagi/commcare-hq/commit/8a55777c09c53642d36d421a6d9143105f23c3e1#diff-b7aa98926b6a7d2a4fd1602b522843de0cbca556cbccbb2017d347f7841905caR201
https://sentry.io/organizations/dimagi/issues/3212087788/?project=136860&query=is%3Aunresolved+domain%3Acrs-remind&statsPeriod=14d

Additionally, there's this disclaimer at the top of the file:

> WARNING this module looks up cases in Couch, but the domain (crs-remind) has
> been migrated to SQL, so any cases created or updated after the migration was
> completed (2021-05-12) will either cause errors or incorrect report output.