Remove CRS custom code
I already removed the main report, and the two remaining views (children
of the report) are both broken and have been for years.

The rows of that report contained links to a separate page, which
doesn’t error outright, though it doesn’t display any data, and it’s
only reachable if you manually create the URL.  As per [this
comment](https://github.com/dimagi/commcare-hq/commit/72573d830102af522c06ee0e50977a2aaa514ddc),
that stopped working for new cases when the domain was migrated to SQL
in [May
2021](https://github.com/dimagi/commcare-hq/commit/ed9f7bb0dabf6807e2b6b44607745999405c9cd6),
and it was later broken for old cases in January of this year (you
cannot access SQL cases as it does in the template)

That page is also “exportable to pdf”, but when I do that I get a blank
html page with a .pdf suffix on the download.  That doesn’t seem to have
ever worked, at least not since 2013.  The line that actually included
data in the response was commented out
[here](https://github.com/dimagi/commcare-hq/commit/7d615f737cc70f62397b6d63503f52d6f0e5aa20#diff-c8e1d3728d48d776e87df61faa5d15097cbaa8633cf86883e970f5f1b62a33ecR40)
and removed
[here](https://github.com/dimagi/commcare-hq/commit/fb693aed99e7dd6f4507701e2f453e896a95c3cd#diff-c8e1d3728d48d776e87df61faa5d15097cbaa8633cf86883e970f5f1b62a33ecL34)
in 2017.