Remove broken and unnecessary cleanup_accounting()
cleanup_accounting() attempts to delete a bunch of SQL rows that will be automatically discarded on test transaction rollback.

However, it never worked because the first thing it tries to do is delete Subscription objects, and Subscription.objects is a DisabledManager, which raises NotImplementedError on get_queryset(), so it always raised an error before anything was deleted.

Error in cleanup_accounting:
  File "corehq/apps/enterprise/tests/utils.py", line 103, in cleanup_accounting
    Subscription.objects.all().delete()
  File ".../site-packages/django/db/models/manager.py", line 156, in all
    return self.get_queryset()
  File "corehq/apps/accounting/models.py", line 1097, in get_queryset
    raise NotImplementedError
NotImplementedError: