Return a new list rather than mutating one
The test changes here are illustrative of the reason why (besides
clarity).  We inserted into a list, mutating it, while iterating over
that same list.  This is why the `list()` coercion in the `for`
statement was necessary - it made a copy of the old version of the list.
However, each time we add new elements to the current position, the `i`
variable drifts behind.  This doesn't affect anything where there's only
one insertion, but these two tests called that insertion twice, so the
second was off-by-one.