Switch from aggregation to python implementation
For reasons unknown to me, the aggregation is very slow, regardless of
the number of docs matching the filters.  This implementation is
substantially faster, by comparison.  My hunch is that this is somehow
analogous to the performance issues we've had sorting by case properties
in this index.  Like, when operating on nested docs, ES isn't able to
pre-filter to those contained in top-level docs matching the query.
Hopefully a more modern version of ES will fix that.