Use simple cache keys
This has been throwing a 500 error when trying to assemble the
quickcache key:
    Bad type "<class 'django.utils.functional.lazy.<locals>.__proxy__'>": []

https://sentry.io/organizations/dimagi/issues/3372326722/
https://sentry.io/organizations/dimagi/issues/3319307493/

That looks to me like something in there is a to-be-translated proxy
string, though I can't figure out what.  The arg to the cache key is
currently an entire case JSON, which seems excessive, since the function
needs only domain and case ID from that JSON.  We lazily include
things like indices and transactions in the "JSON", and quickcache
doesn't know how to serialize thos promises.

This does mean that previously the cache would miss whenever the case
JSON changed, whereas this won't.  The next commit switches this to
cache in the UCR evaluation context.