Change Update Translations PR/commit text to skip CI
This reflects our consensus decision not to run the build on these
automated translations changes, just like we haven't been doing
with our previous process.

In addition to the quirks that prevent CI from running on PRs
opened by a previous GitHub Actions workflow, we decided that the CI
tests did not provide much if any value in catching errors in the
automated translations file, and that such errors are much more likely
to be caught by the previous GA workflow that generates the translations
changes.