Drop unused UCR DATA_TYPE_BOOLEAN
According to commit 065306517b36ab36200e16ce1fd6cc0e5b6245d8, the
boolean data type is used in ES code. But it is not included in
DATA_TYPE_CHOICES, so it has never been possible to create a UCR via the
HQ UI that uses it, and there are no custom static UCRs that use it
either.

Boolean is an obvious data type to support, but DATA_TYPE_BOOLEAN has no
usages in the codebase. We don't seem to have needed it since
2018-06-13, when it was added to the other data types.

(Note: The boolean data type is different from the boolean indicator
type. See corehq.apps.userreports.indicators.factory.IndicatorFactory
for supported indicator types.)