Ignore conflicts on sync to SQL
While the race condition is certainly there, the frequency of this error in production (a couple per day) makes it seem unlikely that concurrent table edits are actually causing the issue and it's not a bug in the uploader itself. However, that bug has thus far avoided detection. Using concurrency with threads in a test was a last resort to show that the bug exists and the fix is effective. In any case ignoring conflicts is fine here because the sole purpose of syncing to SQL is to create parent records that are required for foreign key references by child records.