Fixed logic for parsing date once two-digit year has been converted.
I'm not sure if or how the previous code ever worked: parseInputDate doesn't
give access to the default implementation of parseINputDate, only to the
option the user passed in (if any). The only way I found to access the default
implementation is to call the "private" function, which is questionable, but
better than re-implementing the widget's default parsing.