Move download_reports, make dependent
Move the "Download and Email Reports" permission to the "Reports"
section and make it dependent on the reports permissions.  It will now
only be enabled if the role has access to at least one report and it
hasn't been specifically disabled.

The old behavior was to default this to True always, under the
assumption that all views checking this permission would also check a
report permission.  I think they do, but this feels safer.