Don't auto-enable mobile, api, or reports access
This sets the default role to a minimum permissions set - with only
access_all_locations and report_an_issue enabled by default.  The others
were defaulted to true for ease of rollout when they were added, but now
we can safely reset those so going forward only minimal permissions are
enabled without direct intervention.

The most notable effect of this change is that new roles will not have
mobile endpoint access.  That means if a project assigns mobile workers
to such a role, those users will lose access to the mobile app.  This
can be easily re-enabled, but it must be an explicit decision.  We'll
need to be communicative about this change, as I could see it tripping
up some projects.