Removed old inactivity error
This removes the other usage of isWebApps. The 'Click here to log back in' link
is from https://github.com/dimagi/commcare-hq/pull/27572, an early attempt to handle
auto-logout in web apps, later replaced by the popup-based approach that was
first added in https://github.com/dimagi/commcare-hq/pull/27630, so at this point,
it's fine to remove.