Remove response_body field
I started just turning this into str(response.body) and adding a comment
like, "In the future...", but then I pictured myself looking at that
comment in a year and shaking my head about an imagined future that
never arrived, so I figured YAGNI should apply.  That said, if we're
fairly confident we _will_ want to add a response transform in the not
too distant future and we're confident that we'll log that as a string,
then keeping this in would let us avoid back-populating that field when
we get there.