If you already have Python 3.9, use standard venv
[venv](https://docs.python.org/3/library/venv.html) has been available in the
Python Standard Library since 3.3, and it is the simplest way to manage
virtual environments if you are already using the deadsnakes PPA to manage
Python versions.

(For everyone else, pyenv is the simplest way. Although not applicable to HQ,
pyenv also has more features that venv, and offers Python versions that are
not available in the deadsnakes PPA, like non-CPython versions of Python.)