Attempt to fix intermittent test failure
`TestTeardownExceedsSumOfOtherLimits.test_time_limit2` has been failing recently on Github Actions. A common feature of all observed failures is that the "setup" phase of the test (which ends in `TimingPlugin.startTest`) takes an inexplicably long time. A typical failure includes this unexpected output:

```
  ERROR: runTest2 (corehq.tests.test_nose_timing.TestTeardownExceedsSumOfOtherLimits.makeSuite.<locals>.Test)
    ...
  AssertionError: setup time limit (0.05) exceeded: 0.7901391983032227
```

The theory of this attempted fix is that garbage collection triggered during that "setup" phase is causing the failure. It is a bit strange that the failure has always occurred in the same place. It is unclear why garbage collection would run so consistently. This fix disables garbage collection for the time-critical section of all `TimingPlugin` tests, so it should theoretically prevent garbage collection from interfering with any of them.