Add custom assertions to app settings
This felt cruelly complicated to implement, and not in love with the
result.  I wanted this widget to appear on the "Advanced Settings" tab,
but that is controlled by another knockout model, causing the two to
conflict.  I tried making this a knockout component, which is probably
the right answer, but couldn't get it working.