Remove backup token from user pages
I initially kept it in mobile worker pages, since the ownership model of
those is different, even going so far as to make it display only with an
auditable view, before thinking YAGNI and abandoning that approach.
Preserved for posterity in https://github.com/dimagi/commcare-hq/pull/32950

My rationale being that backup tokens for mobile workers only applies to
projects that have two factor auth, and which have mobile worker roles
that can log on to HQ.  Given that this negatively impacts security,
it's better to not allow it unless we decide it's truly necessary.