Rely on autogenerated instance declarations
Background in https://dimagi-dev.atlassian.net/browse/QA-5216?focusedCommentId=272531
These itemsets are used by case search properties that use lookup table
questions from either a lookup table or a mobile UCR.  This code generated the
instance declaration for those lookups, but there's also other code in this file
that looks up instances used in the nodeset (an xpath expression).  As long as
the nodeset is prefixed with either commcare-report: or item-list: we should be
able to infer the instance to be used without needing to store it directly in
config (references which can become stale when the app is copied).