Record histogram errors
These should show up in datadog as the same metric, with the duration as
"error".  Prometheus expects histogram values to be numbers, so this
doesn't fit in to their system well.  Instead, I'm recording it not as a
counter, which should work with both (though I'm not positive how
Prometheus will handle being sent events as a histogram and as a counter
with the same name)