Use const instead of `function.__name__`
Rather use the const that already exists, than use `function.__name__`
as a const. `DEID_TRANSFORM_FUNCTIONS` shows that these consts are
already being used to look up the transform functions. We should either
stick with that, or we should change the current usage to use
`function.__name__` instead, and the second option looks impractical.