Sort case API by copy of ID for performance reasons
I opted to manipulate the query this way rather than integrate it into
the ESQuery framework as this is a one-off hack and not something we're
likely to be repeating.  Open to second guessing that decision, though.

This sort syntax breaks in ES6, I believe, but this code is covered by
tests, so any breaking changes should surface easily