Ensure duplicate menu instances are merged
This is sort of a recurrance of an issue I introduced in
https://github.com/dimagi/commcare-hq/pull/33037 and then fixed in
https://github.com/dimagi/commcare-hq/pull/33271/ See that second PR
description for details

This was then broken in this commit:
https://github.com/dimagi/commcare-hq/commit/8db8f4a27621307125a7ec627a7b05a2f89f1d6d
The test created in that second PR above demonstrated that instances in
the second of two modules with the same ID will be propagated to all
forms in both modules, but it neglected to verify that same property of
instances required by the first of those two modules.  That latest commit above
had subsequent menus overwriting instances from earlier ones.