Don't fail usercase task on missing users
Sentry flagged this recent PR
https://github.com/dimagi/commcare-hq/pull/33545
as being a likely cause for this issue:
https://dimagi.sentry.io/issues/4497657227/
Turns out it's a longstanding issue that just changed
https://dimagi.sentry.io/issues/3908070098/

I can't figure out exactly what the cause is - looks like something QA
is doing, as all the events I looked at were on a QA domain.  One
likely possibility is that it's just a race condition where users are
saved, but then deleted before the task runs.  Given that we don't run
the task when users are deleted, this seems like an acceptable solution
https://github.com/dimagi/commcare-hq/blob/9f09abf6643780b53eb6ffa74eb8f216fa26d559/corehq/apps/users/models.py#L1748-L1749