check for missing tables in get_adapters instead
the main reason for doing this was to avoid interfering with
RebuildTableTest tests since they are fragile. I don't think it really
matters where we filter out adapters with deleted tables, as long as
we do at some point. The other option is in the actual process_change
and process_changes_chunk methods, but that would mean adding code
to two separate places which I don't like.