Test change profile by name, not ID
Profile names are displayed in the download, and users can reassign
profiles by editing that column.  The tests, however, modify profile by
editing the `commcare_profile` custom data field, so they're effectively
testing something else entirely.  Oddly enough, this does work, as long
as you can figure out the profile ID (it's not shown in the UI), and you
upload it as an integer datatype.

This behavior seems wrong, so I'm explicitly disallowing it