Do not sync attempts on save unless requested
Syncing submodels using the Couch document model is expensive. It would be wildly inefficient to rewrite all attempts every time a repeat record is saved, not to mention when a new attempt is being added. However, proper submodel syncing is needed when copying old documents from Couch to SQL. New attempts are only added in non-migration code in one place, RepeatRecord.add_attempt(), which will be addressed in another commit.