prevent general exceptions from disrupting subsequent processing of records
the changes here stemmed from adding a catch of general exceptions in
addition to the existing ResourceConflict exception, and trying to get
test_subsequent_records_are_still_processed_if_general_exception_encountered
passing. Once general exceptions are handled, the existing logic for
fetching the next record to process does not work, since it will always
fetch the same record that raises an error "...(.order_by('pk')[:1]"

Now the record ids to process are fetched first, and then a lock is
obtained for one record at a time as we process them. This could still
lead to a similar scenario as above, except the first `chunk_size`
records would have to all raise errors which is certainly possible.

We may want to add retry logic here to ensure corrupt records do not
clog this queue.