Test process_reporting_metadata_staging with locked record
This is an inelegant solution to testing behavior with a locked
record without needing to use a TransactionTestCase, which are
much slower than TestCase, and resort to truncating tables to
cleanup the database after the test run which can cause issues
with other tests. Think software plans and other sql objects
that are created as part of initial CommCare HQ setup.