Handle datetime strings in user data
If user data contains a datetime string, jsonobject parses it into a
date object, which we can't send directly to a SQL json field.  To
address this, I use the json-compatible version of the user object and
extract the user_data from that, using jsonobject's serialization.

A side effect of this change is that I cannot pass the object by
reference, meaning updates to it are not passed back to couch.  I think
this is actually an improvement, as the prior behavior was inconsistent
anyways (it'd only pass back when first initialized, and not
thereafter).