Register custom 2FA methods on Django app startup
Replaces the need for the setup_two_factor decorator as the custom
methods are setup at the app level, and the HQTwoFactorMethodForm
controls which methods are available to the user.

The one side effect is that the user encounters the
HQTwoFactorMethodForm regardless of if they are allowed to use phone
methods for 2FA, whereas previously the two factor setup wizard was
smart enough to skip the method step if only one method was registered.