Use as_action where it is possible to have TOTPDevice as other
Initially it is only possible to have phone devices in the
list of other_devices as part of the page context. However,
when a user wants to use their backup phone instead, that
triggers a POST on the accounts/login endpoint which will
generate new context_data, with the TOTPDevice included in
the list of other_devices. This causes an AssertionError
when calling device_action on each device since this is a
phone specific method. Using as_action instead should avoid
the AssertionError.