Replace default_devices with manual fetch of default device
The default_devices method caches the device object dynamically
on the user object passed in. This side effect of what should be
a simple getter requires us to clean up the cached attribute in this
context since we just want to delete the default device. Therefore,
lets just not use it.