Make foreign key test for domain dump pass
CaseDeduplicationActionDefinition and UCRExpression were simply in
the UNKNOWN_MODELS list, so moving those into the dumped models was
straightforward. I double checked that the filters specified for each
model are correct, and tested by running dump_domain_data locally.

SMSCallbackContent appears to no longer be in use, and there is not
an easy way to fetch it by domain anyway. It is easiest to just
confirm if a domain has any entries for this object independently when
performing a domain dump.

Permission is an interesting one. While permissions are created on
fresh install, it seems precarious to rely on the creation to remain
consistent to the point where RolePermissions will always foreign
key to the correct Permission.