Create and use CaseIDFilter in SQL domain dump
For databases with lots of cases, it can be time consuming to fetch all
models for a domain that rely on their foreign key relationship with a
case. For example, CaseTransaction.objects.filter(case__domain=domain).
This requires iterating through all transactions to inspect the case in
the relationship.

Instead, we can fetch all case IDs for a domain with an inexpensive
query since cases are indexed by domain, and use those IDs to fetch the
related models, without ever needing to iterate over the entire table.