Add throttle option to load_domain_data command
Only used by CouchDataLoader for now.

In practice, couch can be easily overloaded by the high load encountered
when saving many different documents. The throttle value of 0.25 seconds
has worked with projects that have encountered this issue. If in the
future this value is insufficient, it might be worth allowing the
calling user to specify a throttle value rather than setting a flag.