Reference latest commit of django-two-factor-auth off of 1.16.0
The change will be part of the next release of django-two-factor-auth,
so once the next release is made we can switch back to referencing
releases directly.

This resolves an issue with the phone call method due to invalid XML in
the token.xml template file used by Twilio when providing the code.