Prefactor: Django>=4.1 added support for lazy objects in mark_safe
See https://docs.djangoproject.com/en/5.0/releases/4.1/#utilities

The code prior to this commit presents an issue on Django>=4.1 because
we manually convert mark_safe to a lazy function and expect it to always
return a str (or SafeString). However since lazy arguments are now
supported, mark_safe might return a __proxy__ object, which our current
code is not designed to handle.