Add reset_encrypted_fields option to serializer
This option ensures that encrypted fields are reset when dumped to a
file. The assumptiom is that the serialized data will be used in a
different environment than the source environment, so the encryption
key will be different in the destination environment. In this case, we
want to reset encrypted fields to make it easier to reset.