Rename to "indexed metadata"
I've been confused about what these properties represent in the past, and have
talked to others who also weren't sure how to use it properly.  I spent way too
much time staring at this in the course of this work and it got a little in my
head, but I think this is an improvement.  Won't be opposed to further fiddling
if you've got better ideas.