Disable pytest warnings recorder
It conflicts with corehq.warnings, which converts all non-whitelisted
warnings into errors in tests. Consider replacing corehq.warnings with
pytest warnings module.

Also remove nose warning from the whitelist.

It may be possible to replace corehq.warnings with pytest's warnings
plugin, but that is not being attempted at this time.