bug: Ensure Tableau config permission is only granted and option shown  when FF is enabled
Prior to this fix, when the FF was disabled and a role was granted web user edit/view permission, the role was inadvertently given Tableau config permission in the background.

With this change, Tableau config permission is only provided when the FF is enabled and explicitly set. This ensures consistency and prevents unintended permission grants.