Updated modal logic
This changes existing behavior based on feedback from QA. The new version modal now supersedes
the warning modal. The new version modal is only shown after HQ is pinged to see if the user is
still logged in, which means the warning modal will always already be showing. So the new logic
is to hide the warning modal if we get a signal to show the new version modal.