Adjusted responsive design for header on case search screen
This broke with https://github.com/dimagi/commcare-hq/pull/34671/, which
updated menu headers to not display on any small screens (as opposed to
removing them specifically on app preview). This particular menu header
contains buttons that do need to be displayed on small screens.
Compromise by hiding just the headeer text, but leaving the buttons.