Give "unlimited" domains 1000 users for rate limit
API access is rate limited based on two scopes, domain, and
subscription.  Access is allowed if either of the two scopes allow
access.  The domain scope is based on the number of mobile workers, and
the account scope is based on the number of mobile workers the plan
allows to be created in a month.

As of this writing, Standard plans get 125 mobile workers per month, Pro
plans get 250, and Advanced get 500.  Enterprise plans get
UNLIMITED_FEATURE_USAGE, which is stored as -1.  The rate limiting code
interprets this value literally, meaning that highest tier plan gets the
lowest level of access.

We don't actually want to remove rate limiting, even for enterprise
plans, so here I'm adding a somewhat arbitrary limit - double the
current level of the plan below it.

This shouldn't decrease any existing project’s effective rate limits,
and would only increase rate limits for projects where the number of
mobile workers is less than this fallback value.

https://dimagi.atlassian.net/browse/SUPPORT-20037