Replace order classes with insert at right spot
In Bootstrap 3 we were able to have and intermediate tag as parent
of the apps. This is not allowed anymore in Bootstrap 5. The order
classes previously used only change the visual order, not the tab
order. Adding a hidden reference div and inserting the apps before
it results in the right order that also keeps the tab order in place.