Select case host if case parent is None
When "The Case's Parent Case" is selected as the messaging
recipient fall back to the case's host if the case has
no parent.

In most instances a case would have either a parent or a host
and not both. This allows the same recipient type to be used
in both instances.