Wait for serverup
Some time after the xsfRequest promise was introduced the code
that awaits it was removed. Looking further into the issue, I
realized the original bug was caused by $.get returning a promise
itself. So instead of creating a wrapping promise we can just wait
for it directly. The best place is right before we call .start.