Await xsrfRequest before starting the app
Instead of awaiting the xsrfRequest before calling
get_endpoint, await it before starting the marionettejs
application. That way we do not need to check the promise
before each request and any other request made by the app
will also have the xsrf token. Not matter which ones is
executed first.