Do not attempt to refetch deleted domain_obj
The domain_obj has a doc_type of "Domain-DUPLICATE" at this point and is
therefore not available in the domain/domains couch view. The object we
have in memory is properly updated and its rev is incremented after
calling domain_obj.save(), so there is no need to try to refetch the
object anyway.