Make query search buttons sticky
Tricky part here was making sure the buttons don't overlap the
breadcrumbs when present, but go to the top of the screen when they
aren't.  I couldn't figure out a good way to do this by structuring the
DOM properly, so I decided to set a class on the containing element
indicating whether the breadcrumbs were present, then use this for the
CSS positioning selector.