Use flexbox for columns in main webapps content
The old behavior was to conditionally offset the "main" section to allow
room for the sidebar, and then put the sidebar on top of that offset
margin, with position: absolute.  This made the sidebar overflow the
div, which messed up the sticky scrolling effect.

We're also planning to soon make the persistent menu and case search map
toggleable, and I think using flex containers will make that all a bit
easier.