Make non-tastypie URLs also available at v1
One hesitation I have about this change is that it modifies the
URL returned by `reverse` for each of these to the new latest version.
That means for example, that those of these which return a 'next'
parameter will have it now pointing to the newer version.  The view is
the same in both instances, so the only way I can imagine it breaking is
if the client has something that checks the format of the URL returned.