Paginate record IDs instead
I don't have a good enough understanding to explain why, but paginating
repeat records by next_check and ID don't work when the table is seeing
lots of writes. The issue was that even if there were 500k records in
the backlog for a partition, closer to only 50k records were being
paginated.