Move and refactor get_repeat_record_ids helper
This should live in the RepeatRecordManager and shouldn't have
validation logic in it. Previously, this helper would return an empty
list if no repeater id or payload id was specified, but given this is
just a database helper, it shouldn't do that. I've added a TODO in the
caller that cares about that to validate the input prior to scheduling
this task to prevent fetching all repeat record ids for a domain.