Reorganize flex divs to make things work better
The main problem was that setting display: flex inside of a container
apparently makes that container's set width more of a suggestion.  This
meant something especially wide inside it might overflow off the page.
Any means of setting a width on the inside element (webforms, in this
case) would address the problem.  This reorganization better reflects
that the main thing being "contained" is either a menu or a webform, and
lets that operate as a normal bootstrap container, while being flexed
as a whole with the sidebar, when present.