Move bottom margin to child container
The `webform-nav` div is always present and visible in webapps, but its
child element, `formnav-container` is only visible in certain contexts.
By attaching the margin to the latter, it doesn't cause the form to be
offset when the element is invisible.  This distinction didn't seem to
be noticeable until I wrapped this all in a flexbox.  Still not sure why
that matters :/