get persistent case tile working with new layout
bottom margin should only be added if persisent case tile is active. Otherwise the bottom margin creates blank space at the top of the form on mobile (and I don't think it should be there when not being used anyway). Could also be fixed with a flexbox but this solution seems simple enough and I think we have enough flexboxes at this point.