Remove negative z-index from shaded elements
This seems to be the cause of
https://dimagi.atlassian.net/browse/USH-5012

It was originally added here:
https://github.com/dimagi/commcare-hq/pull/33908/files#diff-9ba2478f84098bf46ce7d7af2e93881d529b6e0cc6e425d7857a17f0ac992438R264
I asked Robert why it was done that way, and he says "If I remember
right, it's -1 because the shading was covering the text otherwise."

Removing it doesn't appear to cover up the text, so maybe that issue has
been resolved anyways.  Still not sure how/why this only just started
becoming a problem.  My suspicion is that a recent Chrome update
might've caused the behavior change.