Change indexes
After some analyzing, it turns out that the query is between two and
three times as fast using a partial index on RepeatRecord.repeater_id
where state is pending or failed. Performance is also improved using a
partial index with Repeater.is_deleted instead of a normal B-tree
index. When these two indexes are used, the next_attempt_at_partial_idx
is not used.