Treat web user location fields the same
For more context, see https://github.com/dimagi/commcare-hq/pull/34401/
I still _much_ prefer the way I had it with web users, where it's
injected in the usercase and restore code at the last minute, but I
think consistency between the two models is more important still.

The big difference here is that the way I did it for WebUsers made those
fields always present, but sometimes empty.  I do think this is a better
way of doing things, but "missing" and "null" are treated differently on
mobile, so it's conceivable that doing that change for mobile workers
would break someone's app.  I think this is the easiest way to ensure
web and mobile workers behave the same without breaking existing apps