Migrated manage notifications page to bootstrap 5
This is to avoid needing to split the notifications/js/manage_notifications fiel into a B3 and B5 version.
The notifications app will be split between B3 and B5 for the entire bootstrap migration, because
global_notifications.html is part of HQ's base template so it needs both a B3 and B5 version until B3 is totally
gone. But the management page is just an admin page that can be migrated, with the B3 version deleted.